Apply LLVM recommended changes
A number of places in the Solaris code had code paths that would
return garbage due to uninitialised variables. Running through
the cleanup that is provided as part of LLVM enabled a number of
these issues to be closed, as well as identifying situations in which
the Apple ported code could be tightened up as well.